+++ NEWS.md
@@ -1,6 +1,7 @@
 # teal.slice 0.2.0.9001
 
 * Examples now use `scda.2022` instead of `scda.2021`.
+* `FilteredDataset` gains a `get_hash_call()` function to aid reproducibility checks.
 
 ### Enhancements